Split lowering rules out of jax_primitives.py
